+++ Back_End/src/main/java/com/example/sport_full/models/UserModels.java
@@ -3,11 +3,13 @@ package com.example.sport_full.models;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import jakarta.persistence.*;
+import org.hibernate.annotations.Where;
 
 import java.io.Serializable;
 
 @Entity
 @Table(name = "usuarios")
+@Where(clause = "estado_cuenta = false")
 public class UserModels implements Serializable {
 
     @Id
@@ -33,6 +35,8 @@ public class UserModels implements Serializable {
     // Relación opcional con AdminModels
     @OneToOne(mappedBy = "userModels", cascade = CascadeType.ALL, optional = true)
     private AdminModels adminModels;
+
+    private boolean estadoCuenta;
 
     // Getters y Setters
 
@@ -90,5 +94,13 @@ public class UserModels implements Serializable {
 
     public void setAdminModels(AdminModels adminModels) {
         this.adminModels = adminModels;
+    }
+
+    public boolean isEstadoCuenta() {
+        return estadoCuenta;
+    }
+
+    public void setEstadoCuenta(boolean estadoCuenta) {
+        this.estadoCuenta = estadoCuenta;
     }
 }
